+++ knytt/objects/bullets/SuperBullet.cs
@@ -2,7 +2,7 @@ using Godot;
 
 public class SuperBullet : BaseBullet
 {
-    public override void _PhysicsProcess(float delta)
+    public override void _PhysicsProcess(double delta)
     {
         base._PhysicsProcess(delta);
         Translate(new Vector2(GDKnyttDataStore.random.Next(3) - 1, GDKnyttDataStore.random.Next(3) - 1) * delta * 30);
